updated group

+++ server/internal/feature/subject/delete_subject/usecase.go
@@ -2,6 +2,9 @@ package delete_subject
 
 import (
 	"context"
+	"fmt"
+	"server/internal/domain"
+	"server/internal/pkg/usecase"
 
 	"github.com/google/uuid"
 )
@@ -20,6 +23,12 @@ func New(postgres Postgres) *UseCase {
 	}
 }
 
-func (u *UseCase) DeleteSubject(ctx context.Context, id uuid.UUID) error {
+func (u *UseCase) DeleteSubject(ctx context.Context, identity usecase.Identity, id uuid.UUID) error {
+	if identity.Role() != domain.RoleAdmin {
+		return usecase.NewAuthError(
+			fmt.Errorf("user is not admin"),
+		)
+	}
+
 	return u.postgres.DeleteSubject(ctx, id)
 }
